requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
openai
streamlit